Add a fuzzer for testing randomized operations
The fuzzer creates a random sequence of operations and tries them
against a JetKV store. If an inconsistency or panic is detected, the
sequence is tried to be minimized into a minimal reproducible example.
The fuzzer also outputs a seed, which can be used to create the same
sequence of operations again.

Currently only targets the file backend, mostly because it contained
more logic. With some changes, the fuzzer could also target the memory
backend.

No windows support because `fork()` is used, but it shouldn't be a big
change to use `ChildProcess` instead.